refactor(library): extract library toolbar widget
