+++ mobile/lib/screens/library_screen.dart
@@ -471,7 +471,7 @@ class _LibraryViewState extends ConsumerState<_LibraryView>
                   child: Column(
                     children: [
                       if (!widget.selectionMode)
-                        _LibraryToolbar(
+                        LibraryToolbar(
                           isLibrarySelectionMode: isLibrarySelectionMode,
                           canExitSelectionMode: !selectionLockedToCloseOnly,
                           isClipsTabActive: isClipsTabActive,
@@ -559,91 +559,6 @@ class _LibraryViewState extends ConsumerState<_LibraryView>
             ],
           );
         },
-      ),
-    );
-  }
-}
-
-class _LibraryToolbar extends StatelessWidget {
-  const _LibraryToolbar({
-    required this.isLibrarySelectionMode,
-    required this.canExitSelectionMode,
-    required this.isClipsTabActive,
-    required this.onLeadingPressed,
-    required this.onOpenSortMenu,
-    required this.onEnterSelectionMode,
-    required this.onOpenTrash,
-    this.onDeleteSelectedClips,
-  });
-
-  final bool isLibrarySelectionMode;
-  final bool canExitSelectionMode;
-  final bool isClipsTabActive;
-  final VoidCallback onLeadingPressed;
-  final VoidCallback onOpenSortMenu;
-  final VoidCallback onEnterSelectionMode;
-  final VoidCallback onOpenTrash;
-  final VoidCallback? onDeleteSelectedClips;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(16),
-      child: Row(
-        spacing: 8,
-        children: [
-          DivineIconButton(
-            size: .small,
-            type: .secondary,
-            icon: isLibrarySelectionMode ? .x : .caretLeft,
-            semanticLabel: isLibrarySelectionMode && canExitSelectionMode
-                ? context.l10n.commonCancel
-                : null,
-            onPressed: onLeadingPressed,
-          ),
-          Expanded(
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 8),
-              child: Text(
-                context.l10n.profileMyLibraryLabel,
-                style: VineTheme.titleMediumFont(),
-                maxLines: 1,
-                overflow: TextOverflow.ellipsis,
-              ),
-            ),
-          ),
-          if (isClipsTabActive) ...[
-            if (!isLibrarySelectionMode)
-              DivineIconButton(
-                size: .small,
-                type: .secondary,
-                icon: .trash,
-                semanticLabel: context.l10n.libraryTrashEntryLabel,
-                onPressed: onOpenTrash,
-              ),
-            DivineIconButton(
-              size: .small,
-              type: .secondary,
-              icon: .funnelSimple,
-              onPressed: onOpenSortMenu,
-            ),
-            if (!isLibrarySelectionMode)
-              DivineButton(
-                size: .small,
-                type: .secondary,
-                label: context.l10n.librarySelect,
-                onPressed: onEnterSelectionMode,
-              ),
-            if (isLibrarySelectionMode)
-              DivineIconButton(
-                size: .small,
-                type: .error,
-                icon: .trash,
-                semanticLabel: context.l10n.commonDelete,
-                onPressed: onDeleteSelectedClips,
-              ),
-          ],
-        ],
       ),
     );
   }
