+++ mobile/lib/widgets/library/library.dart
@@ -4,4 +4,5 @@
 export 'clips_tab.dart';
 export 'drafts_tab.dart';
 export 'empty_library_state.dart';
+export 'library_toolbar.dart';
 export 'sounds_tab.dart';
